requirements.txt
streamlit
pytesseract
Pillow
langchain
openai

